pseudoramiX: move it under Xext/
