+++ hw/xquartz/meson.build
@@ -76,7 +76,7 @@ libXquartz = static_library('Xquartz',
     srcs_libxquartz,
     c_args: libxquartz_defs,
     objc_args: libxquartz_defs,
-    include_directories: [inc, '../../pseudoramiX', '../../miext/rootless', top_dir_inc],
+    include_directories: [inc, '../../miext/rootless', top_dir_inc],
     dependencies: [xproto_dep, xdmcp_dep, pixman_dep],
     link_with: [libxserver_pseudoramix, libxserver_miext_rootless],
 )
