+++ hw/xquartz/quartz.c
@@ -36,13 +36,13 @@
 #include "dix/dix_priv.h"
 #include "dix/screenint_priv.h"
 #include "miext/extinit_priv.h"
+#include "Xext/pseudoramiX/pseudoramiX.h"
 
 #include "quartzRandR.h"
 #include "inputstr.h"
 #include "quartz.h"
 #include "darwin.h"
 #include "darwinEvents.h"
-#include "pseudoramiX.h"
 #include "extension.h"
 #include "glx_extinit.h"
 #define _APPLEWM_SERVER_
